crypt-wrap-unistd.h: New header file to safely include <unistd.h>.
Some versions of the <unistd.h> header file may declare the
prototype of the crypt() function attributing some parameters
passed into the function *MUST NOT* be NULL.

In unity builds this attribute can instruct several compilers
to delete some of the needed NULL-pointer checks inside of our
code, as we cannot include <unistd.h> in just those source files,
which need some of the prototypes and macro definitions the
<unistd.h> header file provides.

When including <unistd.h> we need to make sure such a prototype
will not be used by the compiler to apply code optimizations,
and thus potential deletion of some needed NULL-pointer checks,
to our implemetation of the crypt() function.  As we use the
preprocessor to rename the prototype, we must include <unistd.h>
within <crypt-port.h>, when compiling a unity build, *before* we
define our macros used for symbol-versioning purposes.

Anyways, using a prototype of the crypt() function, which is
declared with those non-NULL attributes, for compiling external
code, that gets linked against a binary copy of the libxcrypt
library, will not cause any such problems.