Github Actions: Add workflow to run testsuite under memory checkers.
This tests two configurations with everything enabled, one built with
GCC and one with Clang, using both Valgrind and ASan+UBSan.

ASan’s “interceptors” for crypt and crypt_r have a semantic conflict
with libxcrypt, requiring a few tests to be disabled for that build;
see commentary in test/crypt-badargs.c.