Github Actions: run distcheck w/ gcc only, perl 5.14, more criticism
On Travis we were running ‘make distcheck’ only with gcc; I don’t see
a need to run it with clang as well.

Install _all_ of the perlcritic modules that we were installing on
Travis, not just the ones bundled with perlcritic itself.

Use the ‘setup-perl’ action (shogo82148/actions-setup-perl@v1) to run
the distcheck build with Perl 5.14, thus consolidating the distcheck
build with the “oldest supported perl” build that we had on Travis.