Github Actions: Pin specific versions of all packages installed from CPAN.
Right now this is because we need to avoid pulling Perl::Tidy v20210625,
which changes the formatting of the setup block that appears in all of
our scripts:

```
@@ -13,7 +13,7 @@
 use open qw(:utf8);
 no  if $] >= 5.018, warnings => 'experimental::smartmatch';
 no  if $] >= 5.022, warnings => 'experimental::re_strict';
-use if $] >= 5.022, re       => 'strict';
+use if $] >= 5.022, re => 'strict';
```

As both Arch and Debian are still shipping v20210402, it would be
inappropriate to demand that all contributors install the newer version.
In the future, we may have similar problems with other perl packages,
so pin everything.